requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-postgres==5.13.1
apache-airflow-providers-amazon==8.5.0
pandas==2.2.2
sqlalchemy==1.4.49
psycopg2-binary==2.9.9
boto3==1.35.0
pyarrow==17.0.0
python-dotenv==1.0.1
jupyter==1.1.1
redis==5.2.0          # ← 新增
pika==1.3.2           # ← 新增 RabbitMQ 客户端